requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --allow-unsafe requirements-dev.in
#
appdirs==1.4.4            # via black, virtualenv
attrs==20.2.0             # via pytest
black==20.8b1             # via -r requirements-dev.in
cfgv==3.2.0               # via pre-commit
click==7.1.2              # via black, pip-tools
coverage==5.3             # via pytest-cov
decopatch==1.4.8          # via pytest-cases
deepdiff[murmur]==5.0.2   # via -r requirements.txt
distlib==0.3.1            # via virtualenv
factory-boy==3.1.0        # via -r requirements-dev.in
faker==4.14.0             # via factory-boy
filelock==3.0.12          # via virtualenv
flake8==3.8.4             # via -r requirements-dev.in
identify==1.5.6           # via pre-commit
importlib-metadata==2.0.0  # via flake8, pluggy, pre-commit, pytest, virtualenv
iniconfig==1.1.1          # via pytest
isort==5.6.4              # via -r requirements-dev.in
makefun==1.9.4            # via decopatch, pytest-cases
mccabe==0.6.1             # via flake8
mmh3==2.5.1               # via -r requirements.txt, deepdiff
mypy-extensions==0.4.3    # via black, mypy
mypy==0.790               # via -r requirements-dev.in
nodeenv==1.5.0            # via pre-commit
ordered-set==4.0.2        # via -r requirements.txt, deepdiff
packaging==20.4           # via pytest
pathspec==0.8.0           # via black, yamllint
pip-tools==5.3.1          # via -r requirements-dev.in
pluggy==0.13.1            # via pytest
pre-commit==2.7.1         # via -r requirements-dev.in
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8
pydantic==1.6.1           # via -r requirements.txt
pyflakes==2.2.0           # via flake8
pymysql==0.10.1           # via -r requirements.txt
pyparsing==2.4.7          # via packaging
pypika==0.42.1            # via -r requirements.txt
pytest-cases==2.3.0       # via -r requirements-dev.in
pytest-cov==2.10.1        # via -r requirements-dev.in
pytest-pythonpath==0.7.3  # via -r requirements-dev.in
pytest==6.1.1             # via -r requirements-dev.in, pytest-cov, pytest-pythonpath
python-dateutil==2.8.1    # via faker
pyyaml==5.3.1             # via pre-commit, yamllint
regex==2020.10.11         # via black
six==1.15.0               # via packaging, pip-tools, python-dateutil, virtualenv
sqlparse==0.3.1           # via -r requirements.txt, yoyo-migrations
tabulate==0.8.7           # via -r requirements.txt, yoyo-migrations
text-unidecode==1.3       # via faker
toml==0.10.1              # via black, pre-commit, pytest
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.3  # via black, mypy
virtualenv==20.0.34       # via pre-commit
yamllint==1.25.0          # via -r requirements-dev.in
yoyo-migrations==7.2.0    # via -r requirements.txt
zipp==3.3.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
pip==20.2.3               # via pip-tools
setuptools==50.3.1        # via yamllint
